Fixed final failing tests
