+++ tests/ga/test_monitor.py
@@ -172,155 +172,6 @@ class TestMonitor(AgentTestCase, HttpRequestPredicates):
         self.assertEqual(0, len(logger.DEFAULT_LOGGER.periodic_messages), "The monitor thread did not reset the periodic log messages")
 
 
-class TestEventMonitoring(AgentTestCase, HttpRequestPredicates):
-    def setUp(self):
-        AgentTestCase.setUp(self)
-        self.lib_dir = tempfile.mkdtemp()
-        self.event_dir = os.path.join(self.lib_dir, event.EVENTS_DIRECTORY)
-
-        EventLoggerTools.initialize_event_logger(self.event_dir)
-
-    def tearDown(self):
-        fileutil.rm_dirs(self.lib_dir)
-
-    _TEST_EVENT_PROVIDER_ID = "XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX"
-
-    def _create_extension_event(self, # pylint: disable=invalid-name,too-many-arguments
-                               size=0,
-                               name="DummyExtension",
-                               op=WALAEventOperation.Unknown,
-                               is_success=True,
-                               duration=0,
-                               version=CURRENT_VERSION,
-                               message="DummyMessage"):
-        event_data = TestEventMonitoring._get_event_data(name=size if size != 0 else name,
-                op=op,
-                is_success=is_success,
-                duration=duration,
-                version=version,
-                message=random_generator(size) if size != 0 else message)
-        event_file = os.path.join(self.event_dir, "{0}.tld".format(int(time.time() * 1000000)))
-        with open(event_file, 'wb+') as fd: # pylint: disable=invalid-name
-            fd.write(event_data.encode('utf-8'))
-
-    @staticmethod
-    def _get_event_data(duration, is_success, message, name, op, version, eventId=1): # pylint: disable=invalid-name,too-many-arguments
-        event = TelemetryEvent(eventId, TestEventMonitoring._TEST_EVENT_PROVIDER_ID) # pylint: disable=redefined-outer-name
-        event.parameters.append(TelemetryEventParam(GuestAgentExtensionEventsSchema.Name, name))
-        event.parameters.append(TelemetryEventParam(GuestAgentExtensionEventsSchema.Version, str(version)))
-        event.parameters.append(TelemetryEventParam(GuestAgentExtensionEventsSchema.Operation, op))
-        event.parameters.append(TelemetryEventParam(GuestAgentExtensionEventsSchema.OperationSuccess, is_success))
-        event.parameters.append(TelemetryEventParam(GuestAgentExtensionEventsSchema.Message, message))
-        event.parameters.append(TelemetryEventParam(GuestAgentExtensionEventsSchema.Duration, duration))
-
-        data = get_properties(event)
-        return json.dumps(data)
-
-    @patch("azurelinuxagent.common.event.TELEMETRY_EVENT_PROVIDER_ID", _TEST_EVENT_PROVIDER_ID)
-    # @patch("azurelinuxagent.common.protocol.wire.WireClient.send_encoded_event")
-    @patch("azurelinuxagent.common.conf.get_lib_dir")
-    def test_collect_and_send_events(self, mock_lib_dir, *_):
-        mock_lib_dir.return_value = self.lib_dir
-
-        with _create_monitor_handler(enabled_operations=["collect_and_enqueue_events"]) as monitor_handler:
-            self._create_extension_event(message="Message-Test")
-
-            test_mtime = 1000  # epoch time, in ms
-            test_opcodename = datetime.datetime.fromtimestamp(test_mtime).strftime(logger.Logger.LogTimeFormatInUTC)
-            test_eventtid = 42
-            test_eventpid = 24
-            test_taskname = "TEST_TaskName"
-
-            with patch("os.path.getmtime", return_value=test_mtime):
-                with patch('os.getpid', return_value=test_eventpid):
-                    with patch("threading.Thread.ident", new_callable=PropertyMock(return_value=test_eventtid)):
-                        with patch("threading.Thread.getName", return_value=test_taskname):
-                            monitor_handler.run_and_wait()
-
-            # Validating the crafted message by the collect_and_send_events call.
-            self.assertEqual(1, len(monitor_handler.event_list))
-            collected_event = monitor_handler.get_mock_wire_protocol().client.send_encoded_event.call_args[0] # pylint: disable=no-member
-
-            # Some of those expected values come from the mock protocol and imds client set up during test initialization
-            osutil = get_osutil()
-            osversion = u"{0}:{1}-{2}-{3}:{4}".format(platform.system(), DISTRO_NAME, DISTRO_VERSION, DISTRO_CODE_NAME,platform.release())
-
-            sample_message = '<Event id="1"><![CDATA[' \
-                             '<Param Name="Name" Value="DummyExtension" T="mt:wstr" />' \
-                             '<Param Name="Version" Value="{0}" T="mt:wstr" />' \
-                             '<Param Name="Operation" Value="Unknown" T="mt:wstr" />' \
-                             '<Param Name="OperationSuccess" Value="True" T="mt:bool" />' \
-                             '<Param Name="Message" Value="Message-Test" T="mt:wstr" />' \
-                             '<Param Name="Duration" Value="0" T="mt:uint64" />' \
-                             '<Param Name="GAVersion" Value="{1}" T="mt:wstr" />' \
-                             '<Param Name="ContainerId" Value="c6d5526c-5ac2-4200-b6e2-56f2b70c5ab2" T="mt:wstr" />' \
-                             '<Param Name="OpcodeName" Value="{2}" T="mt:wstr" />' \
-                             '<Param Name="EventTid" Value="{3}" T="mt:uint64" />' \
-                             '<Param Name="EventPid" Value="{4}" T="mt:uint64" />' \
-                             '<Param Name="TaskName" Value="{5}" T="mt:wstr" />' \
-                             '<Param Name="KeywordName" Value="" T="mt:wstr" />' \
-                             '<Param Name="ExtensionType" Value="json" T="mt:wstr" />' \
-                             '<Param Name="IsInternal" Value="False" T="mt:bool" />' \
-                             '<Param Name="OSVersion" Value="{6}" T="mt:wstr" />' \
-                             '<Param Name="ExecutionMode" Value="IAAS" T="mt:wstr" />' \
-                             '<Param Name="RAM" Value="{7}" T="mt:uint64" />' \
-                             '<Param Name="Processors" Value="{8}" T="mt:uint64" />' \
-                             '<Param Name="TenantName" Value="db00a7755a5e4e8a8fe4b19bc3b330c3" T="mt:wstr" />' \
-                             '<Param Name="RoleName" Value="MachineRole" T="mt:wstr" />' \
-                             '<Param Name="RoleInstanceName" Value="b61f93d0-e1ed-40b2-b067-22c243233448.MachineRole_IN_0" T="mt:wstr" />' \
-                             '<Param Name="Location" Value="uswest" T="mt:wstr" />' \
-                             '<Param Name="SubscriptionId" Value="AAAAAAAA-BBBB-CCCC-DDDD-EEEEEEEEEEEE" T="mt:wstr" />' \
-                             '<Param Name="ResourceGroupName" Value="test-rg" T="mt:wstr" />' \
-                             '<Param Name="VMId" Value="99999999-8888-7777-6666-555555555555" T="mt:wstr" />' \
-                             '<Param Name="ImageOrigin" Value="2468" T="mt:uint64" />' \
-                             ']]></Event>'.format(AGENT_VERSION, CURRENT_AGENT, test_opcodename, test_eventtid,
-                                                  test_eventpid, test_taskname, osversion, int(osutil.get_total_mem()),
-                                                  osutil.get_processor_cores())
-
-            self.maxDiff = None # pylint: disable=invalid-name
-            self.assertEqual(sample_message.encode('utf-8'), collected_event[1])
-
-    @patch("azurelinuxagent.common.protocol.wire.WireClient.send_encoded_event")
-    @patch("azurelinuxagent.common.conf.get_lib_dir")
-    def test_collect_and_send_events_with_small_events(self, mock_lib_dir, patch_send_event, *_):
-        mock_lib_dir.return_value = self.lib_dir
-
-        with _create_monitor_handler(enabled_operations=["collect_and_send_events"]) as monitor_handler:
-
-            sizes = [15, 15, 15, 15]  # get the powers of 2 - 2**16 is the limit
-
-            for power in sizes:
-                size = 2 ** power
-                self._create_extension_event(size)
-
-            monitor_handler.run_and_wait()
-
-            # The send_event call would be called each time, as we are filling up the buffer up to the brim for each call.
-
-            self.assertEqual(4, patch_send_event.call_count)
-
-    @patch("azurelinuxagent.common.protocol.wire.WireClient.send_encoded_event")
-    @patch("azurelinuxagent.common.conf.get_lib_dir")
-    def test_collect_and_send_events_with_large_events(self, mock_lib_dir, patch_send_event, *_):
-        mock_lib_dir.return_value = self.lib_dir
-
-        with _create_monitor_handler(enabled_operations=["collect_and_send_events"]) as monitor_handler:
-
-            sizes = [17, 17, 17]  # get the powers of 2
-
-            for power in sizes:
-                size = 2 ** power
-                self._create_extension_event(size)
-
-            with patch("azurelinuxagent.common.logger.periodic_warn") as patch_periodic_warn:
-                monitor_handler.run_and_wait()
-
-                self.assertEqual(3, patch_periodic_warn.call_count)
-
-                # The send_event call should never be called as the events are larger than 2**16.
-                self.assertEqual(0, patch_send_event.call_count)
-
-
 @patch('azurelinuxagent.common.osutil.get_osutil')
 @patch('azurelinuxagent.common.protocol.util.get_protocol_util')
 @patch("azurelinuxagent.common.protocol.healthservice.HealthService._report")
